requirements.txt
# TUX Droid AI Control - Dependencies
# =====================================

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Telegram Bot
python-telegram-bot>=20.7

# HTTP Client
httpx>=0.25.0

# Configuration
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Type checking (optional)
mypy>=1.7.0

